Changes at configure() at Controller class. Robin, please have a look at that Throw statement. I Fixed the "deadlock" in while statement in that run@ method at workerThread and now, the thread stops properly, no more NPE issues. Im also working on a workerThreadIterative that will run accordingly to the user input in steps button and a text field to inform how much iterations to run per step , default = 1;
I also made the screen a bit bigger.Im thinking we will need an screen before of that one with two options :
-Iterative Mode
-Full-run Mode

Then the current view.
In case we do that :
For iterative mode we will need to change start to step, and stop to pause.
and a place to change the iterative steps.